added SCALE encoding and proof generation to the JS e2e test file
